refactor(scripts): extract shared walkJoomlaArchive; fix site-wide WAF UA
SonarCloud quality gate: new_duplicated_lines_density was 30.6% — the S2H3
backfill mirrored backfill-bth3-history.ts's ~60-line archive walker verbatim.
Extract it into scripts/lib/joomla-archive-backfill.ts (single `?limit=0` index
fetch + bounded concurrency pool + failure cap) and migrate both S2H3 and BTH3
to call it. (BFMH3 uses a distinct older pattern — left as-is.)

bangkokhash.com's WAF has since started 403ing the bot-style
"compatible; HashTracks-Backfill" UA site-wide (BTH3's archive now blocks it
too), so the helper defaults to a desktop-browser UA — verified live against
both /thursday (228 detail pages) and /siamsunday (136). Dropped the per-script
FETCH_HEADERS duplication.
