+++ scripts/backfill-bth3-history.ts
@@ -17,9 +17,8 @@
 
 import "dotenv/config";
 import { runBackfillScript } from "./lib/backfill-runner";
-import { safeFetch } from "@/adapters/safe-fetch";
+import { walkJoomlaArchive } from "./lib/joomla-archive-backfill";
 import { parseNextRunArticle } from "@/adapters/html-scraper/bangkokhash";
-import type { RawEventData } from "@/adapters/types";
 
 const SOURCE_NAME = "Bangkok Thursday Hash";
 const KENNEL_TIMEZONE = "Asia/Bangkok";
@@ -30,82 +29,17 @@ const BASE_URL = "https://www.bangkokhash.com";
 const INDEX_URL = `${BASE_URL}/thursday/index.php/run-archives-bth3?limit=0`;
 const DETAIL_URL_RE = /\/thursday\/index\.php\/run-archives-bth3\/\d+-run-\d+/g;
 
-const FETCH_HEADERS = {
-  "User-Agent": "Mozilla/5.0 (compatible; HashTracks-Backfill)",
-  Accept: "text/html",
-};
-const FETCH_CONCURRENCY = 4;
-const MAX_TOTAL_FAILURES = 10;
-
-async function fetchText(url: string): Promise<string | null> {
-  const res = await safeFetch(url, { headers: FETCH_HEADERS });
-  if (!res.ok) {
-    console.warn(`  HTTP ${res.status} for ${url}`);
-    return null;
-  }
-  return res.text();
-}
-
-async function discoverDetailUrls(): Promise<string[]> {
-  console.log(`  Fetching index: ${INDEX_URL}`);
-  const html = await fetchText(INDEX_URL);
-  if (!html) throw new Error("Failed to fetch BTH3 archive index");
-  const matches = html.match(DETAIL_URL_RE) ?? [];
-  return [...new Set(matches)]
-    .sort((a, b) => a.localeCompare(b))
-    .map((path) => `${BASE_URL}${path}`);
-}
-
-async function fetchAllArchive(): Promise<RawEventData[]> {
-  const urls = await discoverDetailUrls();
-  console.log(`  Discovered ${urls.length} detail URLs`);
-
-  const events: RawEventData[] = [];
-  let failures = 0;
-  let nextIdx = 0;
-  let processed = 0;
-  let aborted = false;
-  let abortReason = "";
-
-  async function worker(): Promise<void> {
-    while (!aborted) {
-      const i = nextIdx++;
-      if (i >= urls.length) return;
-      const url = urls[i];
-      const html = await fetchText(url);
-      if (!html) {
-        failures++;
-        if (failures >= MAX_TOTAL_FAILURES) {
-          aborted = true;
-          abortReason = `${failures} total fetch failures (limit ${MAX_TOTAL_FAILURES})`;
-        }
-        continue;
-      }
-      const event = parseNextRunArticle(html, KENNEL_TAG, DEFAULT_TIME, url);
-      if (event) {
-        events.push(event);
-      } else {
-        console.warn(`  Skipped (no parse): ${url}`);
-      }
-      processed++;
-      if (processed % 25 === 0) {
-        console.log(`  Progress: ${processed}/${urls.length} fetched, ${events.length} parsed`);
-      }
-    }
-  }
-
-  await Promise.all(Array.from({ length: FETCH_CONCURRENCY }, worker));
-  if (aborted) throw new Error(`Aborted: ${abortReason}`);
-  events.sort((a, b) => a.date.localeCompare(b.date));
-  console.log(`  Final: ${processed}/${urls.length} fetched, ${events.length} parsed`);
-  return events;
-}
-
 runBackfillScript({
   sourceName: SOURCE_NAME,
   kennelTimezone: KENNEL_TIMEZONE,
   label: "Walking BTH3 archive",
-  fetchEvents: fetchAllArchive,
+  fetchEvents: () =>
+    walkJoomlaArchive({
+      baseUrl: BASE_URL,
+      indexUrl: INDEX_URL,
+      detailUrlRe: DETAIL_URL_RE,
+      parse: (html, url) => parseNextRunArticle(html, KENNEL_TAG, DEFAULT_TIME, url),
+    }),
 }).catch((err) => {
   console.error(err);
   process.exit(1);
